refactor: Move register route to authenticated group
This moves the registration route from the guest middleware group to
the authenticated middleware group. This requires the user to be
logged in to access the registration page. Additionally, the "Don't
have an account" link was removed from the login page. This is done
to prevent unauthorized users from creating accounts.